libc_rop/requirements.txt
-i https://mirrors.ustc.edu.cn/pypi/web/simple

pwntools==4.13.1
